made the drop table subscriptions migration conditional

+++ db/migrate/20131119112737_drop_old_subscriptions.rb
@@ -1,10 +1,12 @@
 class DropOldSubscriptions < ActiveRecord::Migration
   def up
-  	drop_table :subscriptions
+  	unless table_exists?(:subscriptions)
+	  	drop_table :subscriptions
+		end
   end
 
 	def down
-  	raise ActiveRecord::IrreversibleMigration
-  end
+	  	raise ActiveRecord::IrreversibleMigration
+	end
 
 end
